Moved the recycler in the services folder
